+++ ported/ghost-headline/config.yaml
@@ -1,16 +1,17 @@
 THEME_NAME: Headline
 THEME_VERSION: 1.0.0
+THEME_FONTS: inter:400,500,600,700|lora:400,400i,500,600,700,700i
 POSTS_PER_PAGINATION: 6
 
 colors: 
   light: 
-    accent_color: f51919
-    primary_text_color: 15171a
-    secondary_text_color: 999
+    accent_color: "#f51919"
+    primary_text_color: "#15171a"
+    secondary_text_color: "#999999"
   dark: 
-    accent_color: f51919
-    primary_text_color: f6f6f6
-    secondary_text_color: 999
+    accent_color: "#f51919"
+    primary_text_color: "#f6f6f6"
+    secondary_text_color: "#999999"
 
 fonts: 
   body: 
@@ -28,22 +29,22 @@ settings:
   home:
     tag_1:
       show: true
-      slug: autem
+      slug: welcome
     tag_2:
       show: true
-      slug: deserunt
+      slug: quidem
     tag_3:
       show: true
-      slug: enim
+      slug: omnis
     tag_4:
       show: true
-      slug: laboriosam
+      slug: molestias
     tag_5:
       show: true
-      slug: nulla
+      slug: dolorem
     tag_6: 
       show: true
-      slug: welcome
+      slug: commodi
   loop:
     authors: true
     featured_image: true
