+++ ported/ghost-headline/lang/en.yaml
@@ -7,3 +7,4 @@ close: Close
 not_found: Page not found
 go_home: Go to the front page →
 translations: Translations
+featured: Featured
