+++ ported/ghost-headline/templates/404.twig
@@ -4,9 +4,9 @@
 
 <main id="main">
 
-    <div class="gh-app">
-        <div class="gh-viewport">
-            <div class="gh-view">
+    <div class="hy-app">
+        <div class="hy-viewport">
+            <div class="hy-view">
                 <section class="error-content error-404 js-error-container">
                     <section class="error-details">
                         <section class="error-message">
