+++ ported/ghost-headline/templates/_card-small.twig
@@ -1,5 +1,5 @@
-<article class="gh-card post tag-getting-started">
-    <a class="gh-card-link" href="{{ post.url }}">
-        <h3 class="gh-card-title">{{ post.title }}</h3>
+<article class="hy-card post tag-getting-started">
+    <a class="hy-card-link" href="{{ post.url }}">
+        <h3 class="hy-card-title">{{ post.title }}</h3>
     </a>
 </article>
