+++ ported/ghost-headline/templates/_footer.twig
@@ -1,10 +1,10 @@
-<footer class="gh-foot gh-outer">
-    <div class="gh-foot-inner gh-inner">
+<footer class="hy-foot hy-outer">
+    <div class="hy-foot-inner hy-inner">
 
         {% include '_nav-footer.twig' %}
 
         {% if _config.settings.social %}
-        <div class="gh-social-icons">
+        <div class="hy-social-icons">
             {% include '_social.twig' with {social: _blog.social} %}
             {% import "_macro.twig" as macro %}
             {% if _config.settings.feed %}
@@ -14,7 +14,7 @@
         {% endif %}
 
         {% if _config.settings.copyright.show_hide %}
-        <div class="gh-copyright">{{ _config.settings.copyright.copy_text }}</div>
+        <div class="hy-copyright">{{ _config.settings.copyright.copy_text }}</div>
         {% endif %}
     </div>
 </footer>
