+++ ported/ghost-headline/templates/_nav.twig
@@ -1,4 +1,4 @@
-<nav class="gh-head-menu">
+<nav class="hy-head-menu">
     <ul class="nav"
     {% for nav in _blog.nav_header %}
         <li class="nav-{{ nav.name | lower | slug }}"><a href="{{ nav.url }}">{{ nav.name }}</a></li>
