+++ ported/ghost-headline/templates/_pagination.twig
@@ -1,6 +1,6 @@
 
-<footer class="gh-topic-footer">
+<footer class="hy-topic-footer">
     {% if _pagination.page_next %}
-        <a class="gh-topic-link" href="{{ _pagination.page_next | pagination_page_url }}">{{ 'show more' | lang }} <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><line x1="5" y1="12" x2="19" y2="12"></line><polyline points="12 5 19 12 12 19"></polyline></svg></a>
+        <a class="hy-topic-link" href="{{ _pagination.page_next | pagination_page_url }}">{{ 'show more' | lang }} <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><line x1="5" y1="12" x2="19" y2="12"></line><polyline points="12 5 19 12 12 19"></polyline></svg></a>
     {% endif %}
 </footer>
