+++ ported/ghost-headline/templates/_search.twig
@@ -6,7 +6,7 @@
           <form class="search-form">
             <i aria-hidden="true" class="search-input-icon">{{ icon('ionicons', 'search') }}</i>
             <input class="search-field" type="text" placeholder="{{ 'input_search' | lang }}…">
-            <button class="search-button" type="submit">
+            <button class="search-button">
               <i aria-hidden="true">{{ icon('ionicons', 'close') }}</i>
             </button>
           </form>
