+++ ported/ghost-headline/templates/page.twig
@@ -2,15 +2,15 @@
 
 {% block body %}
 
-<main class="page-template gh-main">
+<main class="page-template hy-main">
 
-<article class="gh-article post no-image">
+<article class="hy-article post no-image">
 
-        <header class="gh-article-header gh-canvas">
-            <h1 class="gh-article-title">{{ _post.title }}</h1>
+        <header class="hy-article-header hy-canvas">
+            <h1 class="hy-article-title">{{ _post.title }}</h1>
         </header>
 
-        <section class="gh-content gh-canvas">
+        <section class="hy-content hy-canvas">
         {{ _post.content | raw }}
         </section>
