+++ ported/ghost-headline/templates/tag.twig
@@ -2,13 +2,13 @@
 
 {% block body %}
 
-	<main class="gh-main gh-outer">
-		<div class="gh-inner">
-			<section class="gh-pagehead">
-				<h1 class="gh-pagehead-title">{{ _tag.name }}</h1>
+	<main class="hy-main hy-outer">
+		<div class="hy-inner">
+			<section class="hy-pagehead">
+				<h1 class="hy-pagehead-title">{{ _tag.name }}</h1>
 			</section>
-			<div class="gh-topic gh-topic-grid">
-				<div class="gh-topic-content gh-feed">
+			<div class="hy-topic hy-topic-grid">
+				<div class="hy-topic-content hy-feed">
                     {%  include '_loop.twig' %}
 				</div>
 			</div>
